+++ lib/lightning_web/live/project_live/settings.ex
@@ -10,16 +10,13 @@ defmodule LightningWeb.ProjectLive.Settings do
   alias Lightning.Policies.Permissions
   alias Lightning.Policies.ProjectUsers
   alias Lightning.Projects
-  alias Lightning.Projects.Project
   alias Lightning.Projects.ProjectUser
   alias Lightning.VersionControl
-  alias Lightning.VersionControl.GithubError
   alias Lightning.WebhookAuthMethods
   alias Lightning.Workflows.WebhookAuthMethod
   alias LightningWeb.Components.Form
-  alias LightningWeb.Components.NewInputs
+  alias LightningWeb.Components.GithubComponents
   alias LightningWeb.LiveHelpers
-  alias LightningWeb.ProjectLive.DeleteConnectionModal
 
   require Logger
 
@@ -28,6 +25,10 @@ defmodule LightningWeb.ProjectLive.Settings do
   @impl true
   def mount(_params, _session, socket) do
     %{project: project, current_user: current_user} = socket.assigns
+
+    if connected?(socket) do
+      VersionControl.subscribe(current_user)
+    end
 
     project_users =
       Projects.get_project_users!(project.id)
@@ -112,12 +113,7 @@ defmodule LightningWeb.ProjectLive.Settings do
         project_user
       )
 
-    {show_github_setup, show_repo_setup, show_sync_button,
-     project_repo_connection} = repo_settings(project)
-
-    if show_repo_setup and connected?(socket) do
-      collect_project_repo_connections(socket.assigns.project.id)
-    end
+    repo_connection = VersionControl.get_repo_connection_for_project(project.id)
 
     {:ok,
      socket
@@ -135,52 +131,13 @@ defmodule LightningWeb.ProjectLive.Settings do
        can_edit_data_retention: can_edit_data_retention,
        can_write_webhook_auth_method: can_write_webhook_auth_method,
        can_create_project_credential: can_create_project_credential,
-       show_github_setup: show_github_setup,
-       show_repo_setup: show_repo_setup,
-       show_sync_button: show_sync_button,
-       project_repo_connection: project_repo_connection,
-       repos: [],
-       branches: [],
-       loading_branches: false,
+       project_repo_connection: repo_connection,
        github_enabled: VersionControl.github_enabled?(),
        can_install_github: can_write_github_connection,
        can_initiate_github_sync: can_initiate_github_sync,
-       pending_github_installation:
-         VersionControl.get_pending_user_installation(
-           socket.assigns.current_user.id
-         ),
        selected_credential_type: nil,
        show_collaborators_modal: false
      )}
-  end
-
-  defp repo_settings(%Project{id: project_id}) do
-    repo_connection = VersionControl.get_repo_connection(project_id)
-
-    project_repo_connection = %{"repo" => nil, "branch" => nil}
-
-    # {show_github_setup, show_repo_setup, show_sync_button}
-    case repo_connection do
-      nil ->
-        {true, false, false, project_repo_connection}
-
-      %{repo: nil} ->
-        {false, true, false, project_repo_connection}
-
-      %{repo: r, branch: b, github_installation_id: g} ->
-        {false, true, true,
-         %{"repo" => r, "branch" => b, "github_installation_id" => g}}
-    end
-  end
-
-  # we should only run this if repo setting is pending
-  defp collect_project_repo_connections(project_id) do
-    pid = self()
-
-    Task.start(fn ->
-      resp = VersionControl.fetch_installation_repos(project_id)
-      send(pid, {:repos_fetched, resp})
-    end)
   end
 
   defp can_edit_digest_alert(
@@ -383,150 +340,6 @@ defmodule LightningWeb.ProjectLive.Settings do
     end
   end
 
-  def handle_event("install_app", _, socket) do
-    user_id = socket.assigns.current_user.id
-    project_id = socket.assigns.project.id
-
-    case Application.get_env(:lightning, :github_app) |> Map.new() do
-      %{app_name: nil} ->
-        Logger.error("GitHub App Name not configured")
-        # Send to sentry and show cozy error
-        error =
-          GithubError.misconfigured("GitHub App Name Misconfigured", %{
-            app_name: nil
-          })
-
-        Sentry.capture_exception(error)
-
-        {:noreply,
-         socket
-         |> put_flash(
-           :error,
-           "Sorry, it seems that the GitHub App Name has not been properly configured for this instance of Lighting. Please contact the instance administrator"
-         )}
-
-      %{app_name: app_name} ->
-        {:ok, _connection} =
-          VersionControl.create_github_connection(%{
-            user_id: user_id,
-            project_id: project_id
-          })
-
-        {:noreply,
-         redirect(socket,
-           external: "https://github.com/apps/#{app_name}"
-         )}
-    end
-  end
-
-  def handle_event("reinstall_app", _, socket) do
-    user_id = socket.assigns.current_user.id
-    project_id = socket.assigns.project.id
-
-    case Application.get_env(:lightning, :github_app) |> Map.new() do
-      %{app_name: nil} ->
-        error =
-          GithubError.misconfigured("GitHub App Name Misconfigured", %{
-            app_name: nil
-          })
-
-        Sentry.capture_exception(error)
-
-        {:noreply,
-         socket
-         |> put_flash(
-           :error,
-           "Sorry, it seems that the GitHub App Name has not been properly configured for this instance of Lighting. Please contact the instance administrator"
-         )}
-
-      %{app_name: app_name} ->
-        {:ok, _} = VersionControl.remove_github_connection(project_id)
-
-        {:ok, _connection} =
-          VersionControl.create_github_connection(%{
-            user_id: user_id,
-            project_id: project_id
-          })
-
-        {:noreply,
-         redirect(socket, external: "https://github.com/apps/#{app_name}")}
-    end
-  end
-
-  def handle_event("delete_repo_connection", _, socket) do
-    project_id = socket.assigns.project.id
-
-    {:ok, _} = VersionControl.remove_github_connection(project_id)
-
-    {:noreply,
-     socket
-     |> assign(show_github_setup: true, show_sync_button: false)
-     |> push_patch(to: ~p"/projects/#{project_id}/settings#vcs")}
-  end
-
-  def handle_event("save_repo", %{"repo" => repo, "branch" => branch}, socket) do
-    if socket.assigns.can_install_github do
-      {:noreply, connect_github_repo(socket, repo, branch)}
-    else
-      {:noreply,
-       put_flash(socket, :error, "You are not authorized to perform this action")}
-    end
-  end
-
-  def handle_event(
-        "initiate_sync",
-        params,
-        %{assigns: %{current_user: u}} = socket
-      ) do
-    if socket.assigns.can_initiate_github_sync do
-      case VersionControl.initiate_sync(params["id"], u.email) do
-        {:ok, :fired} ->
-          {:noreply, socket |> put_flash(:info, "Sync Initialized")}
-
-        _err ->
-          # we should log or instrument this situation
-          {:noreply, socket |> put_flash(:error, "Sync Error")}
-      end
-    else
-      {:noreply, socket |> put_flash(:error, "Viewers Cannot Initiate Sync")}
-    end
-  end
-
-  def handle_event("repo_selected", params, socket) do
-    pid = self()
-
-    Task.start(fn ->
-      {:ok, branches} =
-        VersionControl.fetch_repo_branches(
-          socket.assigns.project.id,
-          params["repo"]
-        )
-
-      send(pid, {:branches_fetched, branches})
-    end)
-
-    {:noreply,
-     socket
-     |> assign(
-       loading_branches: true,
-       project_repo_connection: %{
-         socket.assigns.project_repo_connection
-         | "repo" => params["repo"]
-       }
-     )}
-  end
-
-  def handle_event("branch_selected", params, socket) do
-    {:noreply,
-     socket
-     |> assign(
-       project_repo_connection: %{
-         socket.assigns.project_repo_connection
-         | "branch" => params["branch"]
-       }
-     )}
-  end
-
   @impl true
   def handle_info(:collaborators_updated, socket) do
     project_users =
@@ -538,50 +351,37 @@ defmodule LightningWeb.ProjectLive.Settings do
      |> put_flash(:info, "Collaborators updated successfully!")}
   end
 
-  def handle_info({:branches_fetched, branches_result}, socket) do
-    case branches_result do
-      {:error, error} ->
-        {:noreply, socket |> put_flash(:error, error_message(error))}
-
-      branches ->
-        {:noreply, socket |> assign(loading_branches: false, branches: branches)}
-    end
-  end
-
   def handle_info({:forward, mod, opts}, socket) do
     send_update(mod, opts)
     {:noreply, socket}
   end
 
-  def handle_info({:repos_fetched, result}, socket) do
-    case result do
-      {:error, error} ->
-        {:noreply, socket |> put_flash(:error, error_message(error))}
+  def handle_info(
+        %Lightning.VersionControl.Events.OauthTokenAdded{},
+        socket
+      ) do
+    {:noreply,
+     socket
+     |> put_flash(:info, "Github account linked successfully")
+     |> push_navigate(to: ~p"/projects/#{socket.assigns.project}/settings#vcs")}
+  end
 
-      {:ok, [_ | _] = repos} ->
-        {:noreply, socket |> assign(repos: repos)}
-
-      # while it's possible to trigger this state when testing
-      # GitHub makes it pretty impossible to arrive here
-      _ ->
-        {:noreply, socket}
-    end
+  def handle_info(
+        %Lightning.VersionControl.Events.OauthTokenFailed{},
+        socket
+      ) do
+    {:noreply,
+     socket
+     |> put_flash(
+       :error,
+       "Oops! Github account failed to link. Please try again"
+     )}
   end
 
   # catch all callback. Needed for tests because of Swoosh emails in tests
   def handle_info(msg, socket) do
     Logger.debug("Received unknown message: #{inspect(msg)}")
     {:noreply, socket}
-  end
-
-  defp error_message(error) do
-    case error do
-      %{code: :installation_not_found} ->
-        "Sorry, it seems that the GitHub App ID has not been properly configured for this instance of Lightning. Please contact the instance administrator"
-
-      %{code: :invalid_certificate} ->
-        "Sorry, it seems that the GitHub cert has not been properly configured for this instance of Lightning. Please contact the instance administrator"
-    end
   end
 
   defp dispatch_flash(change_result, socket) do
@@ -713,56 +513,6 @@ defmodule LightningWeb.ProjectLive.Settings do
     """
   end
 
-  defp has_pending_installation?(%{id: project_id}, %{
-         project_id: installation_project_id
-       }) do
-    project_id != installation_project_id
-  end
-
-  defp has_pending_installation?(_project, _installation) do
-    false
-  end
-
-  defp install_github_tooltip(project, can_install_github, pending_installation) do
-    case {can_install_github,
-          has_pending_installation?(project, pending_installation)} do
-      {false, _} ->
-        "You're not allowed to configure repository connections for this project."
-
-      {_, true} ->
-        """
-        You have a pending github installation in another project.
-        <a href="#{~p"/projects/#{pending_installation.project_id}/settings#vcs"}" class="underline text-blue-400">Click here</a> to complete it / remove it before proceeding
-        """
-
-      _other ->
-        nil
-    end
-  end
-
-  defp connect_github_repo(socket, repo, branch) do
-    case VersionControl.connect_github_repo(
-           socket.assigns.project.id,
-           repo,
-           branch
-         ) do
-      {:ok, %{github_installation_id: github_installation_id}} ->
-        socket
-        |> assign(
-          show_repo_setup: false,
-          show_sync_button: true,
-          project_repo_connection: %{
-            "branch" => branch,
-            "repo" => repo,
-            "github_installation_id" => github_installation_id
-          }
-        )
-
-      {:error, _error} ->
-        put_flash(socket, :error, "Oops! Error connecting to github")
-    end
-  end
-
   defp confirm_user_removal_modal(assigns) do
     ~H"""
     <.modal id={@id} width="max-w-md">
@@ -832,5 +582,9 @@ defmodule LightningWeb.ProjectLive.Settings do
   defp user_removable?(project_user, current_user, can_remove_project_user) do
     can_remove_project_user and project_user.role != :owner and
       project_user.user_id != current_user.id
+  end
+
+  defp user_has_valid_oauth_token(user) do
+    VersionControl.oauth_token_valid?(user.github_oauth_token)
   end
 end